+++ src/widgets/teamroom/tool/ui/vote-tool/index.tsx
@@ -2,12 +2,12 @@ import { useMemo, useState } from 'react';
 import { useParams } from 'react-router-dom';
 
 import { TOOL_CONTENTS } from '@/entities/teamroom/setup/tool/model/tool-contents';
-import { useSubmitAllToolVotes } from '@/entities/tool/hooks/useTool';
+import { useSubmitAllToolVotes } from '@/entities/teamroom/tool/hooks/useTool';
 import BottomButton from '@/shared/ui/button/BottomButton';
 import TopBar from '@/shared/ui/header/TopBar';
 import { useModalStore } from '@/shared/ui/modal/model/modal-store';
-import StepTitle from '@/shared/ui/tab/StepTitle';
 import ToolItems from '@/widgets/teamroom/tool/ui/vote-tool/ui/ToolItems';
+import StepTitle from '@/widgets/vote/ui/StepHeader';
 
 interface Props {
   onFinish: () => void;
